+++ app/src/main/java/com/sephoratest/app/navigation/products/ProductNavigationItem.kt
@@ -1,0 +1,9 @@
+package com.sephoratest.app.navigation.products
+
+sealed class ProductNavigationItem(
+    val route: String
+) {
+
+    object ProductList : ProductNavigationItem(route = "product-list")
+    object ProductDetail : ProductNavigationItem(route = "product-detail")
+}