+++ app/src/main/java/com/sephoratest/app/products/list/LikeIconStyleFactory.kt
@@ -1,0 +1,9 @@
+package com.sephoratest.app.products.list
+
+object LikeIconStyleFactory {
+
+    fun fromState(liked: Boolean) = when(liked) {
+        true -> LikeIconStyle.LikedStyle
+        else -> LikeIconStyle.UnlikedStyle
+    }
+}