+++ app/src/main/java/com/sephoratest/app/products/list/ProductsScreen.kt
@@ -1,4 +1,4 @@
-package com.sephoratest.app.products
+package com.sephoratest.app.products.list
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,14 +9,16 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
-import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.swiperefresh.SwipeRefresh
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
+import com.sephoratest.app.ui.models.UiProduct
 
 @Composable
-fun ProductsScreen() {
-
-    val viewModel: ProductsViewModel = viewModel()
+fun ProductsScreen(
+    onProductClicked: (UiProduct) -> Unit
+) {
+    val viewModel: ProductsViewModel = hiltViewModel()
 
     val isLoading by viewModel.isLoading.observeAsState()
     val products by viewModel.products.observeAsState()
@@ -34,9 +36,12 @@ fun ProductsScreen() {
             onRefresh = { viewModel.loadAll() }) {
 
             LazyColumn {
-                products?.let {
-                    items(it) {
-                        ProductListItem(it)
+                products?.let { list ->
+                    items(list) {
+                        ProductListItem(
+                            product = it,
+                            onProductClicked = onProductClicked
+                        )
                     }
                 }
             }