+++ ui/src/main/java/com/sephoratest/ui/atoms/TextTitle.kt
@@ -2,20 +2,26 @@ package com.sephoratest.ui.atoms
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
 
 @Composable
 fun TextTitle(
+    modifier: Modifier = Modifier,
     text: String = "Title",
     color: Color = Color.Black,
     fontSize: Int = 16,
+    textAlign: TextAlign = TextAlign.Start,
 ) {
     Text(
+        modifier = modifier,
         text = text,
         color = color,
         fontSize = fontSize.sp,
         fontWeight = FontWeight.Bold,
+        textAlign = textAlign,
     )
 }