+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added dagger hilt as dependency injection manager
 - Plugged all layers together and initiated compose screen
 - Initiated UI module to contain atomic design of the app and created the product list item
+- Implemented a simple detail screen
 
 ### Changed